server/image_viewer/pyproject.toml
[project]
name = "image-viewer"
version = "0.1.0"
description = "A simple image viewer for images captured by app.py"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "gunicorn>=22.0.0",
    "asyncio>=3.4.3",
]

# [build-system] セクションを削除

[dependency-groups]
# uv run just dev
dev = [
    "just-bin>=1.40.0",
]
